⚙️ FEATURE-#21: Rename email_profile/factories.py to credentials.py
